% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplus_bound.R
\name{nplus_bound}
\alias{nplus_bound}
\title{Shortcut for computing the lower and upper bounds for n+ by using the partitioning principle}
\usage{
nplus_bound(p_right_tail, method = "Fisher", alpha = 0.05)
}
\arguments{
\item{p_right_tail}{vector of p-values for the null hypotheses H_i: theta_i <= 0, i = 1,...,n}

\item{method}{combining function specification. Default is Fisher's combination method}

\item{alpha}{The significance level of the test procedure. Default is 0.05}
}
\value{
(1-alpha)-confidence lower and upper bounds for n+
}
\description{
Shortcut for computing the lower and upper bounds for n+ by using the partitioning principle
}
\examples{
# generate right-tailed p-values
set.seed(123)
x <- rnorm(50, mean = c(rep(2, 25), rep(-2, 25)))
p <- pnorm(x, lower.tail = FALSE)
res <- nplus_bound(p)
res
}
